minor changes are done
