+++ .env
@@ -5,7 +5,7 @@
 # See the documentation for all the connection string options: https://pris.ly/d/connection-strings
 
 # DATABASE_URL="postgresql://johndoe:randompassword@localhost:5432/mydb?schema=public"
-DATABASE_URL="mysql://root:@localhost:3306/OscuroBok"
+DATABASE_URL="mysql://root:password@localhost:3306/OscuroBok"
 SECRET=dfJVFYJfvkuv1546
 
 EMAIL=growthh.in@gmail.com
