+++ README.md
@@ -52,9 +52,10 @@ graph TD
 * git clone https://github.com/OscuroBok/OscuroBok_dashboard_FE.git
 
 # Commands used:
-* npm install
-* npm i --save-dev prisma@latest & npm i @prisma/client@latest (Compulsory**)
-* npm run db:migrate_all ->  Migrates all Schemas and creates tables in Database
-* npm run seed:role (if u don't see the message, "The seed command has been executed") -> This command will insert the Admins roles into DB
-* npm run seed:admin(Inserts the Super Admin credentials and data in DB)
-* npm start
+* npm install (This installs all required dependencies listed in the package.json.)
+* npm run prisma:update (Updates Prisma and generates the latest schema.)
+* npm run db:migrate_all (Migrates all schemas and creates necessary tables in the database.)
+* npm run seed:role (Inserts admin roles into the database. Run this if you don't see the message: "The seed command has been executed.")
+* npm run seed:admin (Inserts both restaurant and admin details into the database.)
+* npm run seed:all (Inserts both restaurant and admin details into the database.)
+* npm start (Starts the application.)
